All jtt tests can be compiled

+++ C1X/src/com/sun/c1x/ci/RiRuntime.java
@@ -191,11 +191,11 @@ public interface RiRuntime {
 
     int interpreterFrameMonitorSize();
 
-    long basicObjectLockSize();
+    int basicObjectLockSize();
 
-    long basicObjectLockOffsetInBytes();
+    int basicObjectLockOffsetInBytes();
 
-    long basicObjectObjOffsetInBytes();
+    int basicObjectObjOffsetInBytes();
 
     int basicLockDisplacedHeaderOffsetInBytes();
